+++ src/main/java/com/vetmanagement/business/abstracts/IAvailableDateService.java
@@ -4,7 +4,7 @@ import com.vetmanagement.entities.AvailableDate;
 import org.springframework.data.domain.Page;
 
 public interface IAvailableDateService {
-    AvailableDate get(int id);
+    AvailableDate get(Long id);
 
     Page<AvailableDate> cursor(int page, int pageSize);
 
@@ -12,5 +12,5 @@ public interface IAvailableDateService {
 
     AvailableDate update(AvailableDate availableDate);
 
-    boolean delete(int id);
+    boolean delete(Long id);
 }
