+++ src/main/java/com/vetmanagement/business/concretes/CustomerManager.java
@@ -20,7 +20,7 @@ public class CustomerManager implements ICustomerService {
     }
 
     @Override
-    public Customer get(int id) {
+    public Customer get(Long id) {
         return this.customerRepo.findById(id).orElseThrow(() -> new NotFoundException(Msg.NOT_FOUND));
     }
 
@@ -41,7 +41,7 @@ public class CustomerManager implements ICustomerService {
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(Long id) {
         Customer customer = this.get(id);
         this.customerRepo.delete(customer);
         return true;
