+++ src/main/java/com/vetmanagement/dao/AppointmentRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AppointmentRepo extends JpaRepository<Appointment, Integer> {
+public interface AppointmentRepo extends JpaRepository<Appointment, Long> {
 }
