+++ src/main/java/com/vetmanagement/dao/AvailableDateRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvailableDateRepo  extends JpaRepository<AvailableDate, Integer> {
+public interface AvailableDateRepo  extends JpaRepository<AvailableDate, Long> {
 }
