+++ src/main/java/com/vetmanagement/dto/response/AvailableDateResponse.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AvailableDateResponse {
-    private int id;
+    private Long id;
     private LocalDate availableDate;
-    private int doctorId;
+    private Long doctorId;
 }
